Introduce InverseKinematicsToolBase and move common properties to it, should be transparent to clients
